refactor: Renamed ColorRampSelector > ColorRampDropdown, cleanup for ids vs classname
